// frontend/src/api/model/memoCommentRead.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Qualis API
 * OpenAPI spec version: 0.1.0
 */
import type { MemoCommentReadUserId } from './memoCommentReadUserId';
import type { MemoCommentReadResolvedAt } from './memoCommentReadResolvedAt';
import type { MemoCommentReadResolvedBy } from './memoCommentReadResolvedBy';

export interface MemoCommentRead {
    id: number;
    entry_id: number;
    user_id: MemoCommentReadUserId;
    body: string;
    mentions: number[];
    resolved: boolean;
    resolved_at: MemoCommentReadResolvedAt;
    resolved_by: MemoCommentReadResolvedBy;
    deleted: boolean;
    created_at: string;
    updated_at: string;
}