// frontend/src/api/model/memoRead.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Qualis API
 * OpenAPI spec version: 0.1.0
 */
import type { MemoReadParentType } from './memoReadParentType';
import type { MemoEntryRead } from './memoEntryRead';

export interface MemoRead {
    parent_type: MemoReadParentType;
    parent_id: number;
    entries: MemoEntryRead[];
}